Create ExternalEntity model
The frontend should not constantly call the CRIS API.
This would otherwise call the 429 error (too many requests).

We will create an ExternalEntity model.
Show Projects including the current cris

Finally: we can create these ExternalEntitys when the user goes to the transform page.

Later, we will connect the ExternalEntity with the Diff…